refactor(workspaces): colocate locations and references
